+++ src/compas_fab/robots/robot.py
@@ -680,41 +680,7 @@ class Robot(object):
         frame_WCF = frame_RCF.transformed(self.transformation_RCF_WCF(group))
         return frame_WCF
 
-    def to_tool0_frame(self, frame_tcf):
-        """Converts a frame at the robot's tool tip (tcf frame) to a frame at the robot's flange (tool0 frame) using the attached tool.
-
-        Parameters
-        ----------
-        frame_tcf : :class:`Frame`
-            A frame (in WCF) at the robot's tool tip (tcf).
-
-        Returns
-        -------
-        :class:`Frame`
-            A frame (in WCF) at the robot's flange (tool0).
-
-        Raises
-        ------
-        Exception
-            If the end effector is not set.
-
-        Examples
-        --------
-        >>> mesh = Mesh.from_stl(compas_fab.get('planning_scene/cone.stl'))
-        >>> frame = Frame([0.14, 0, 0], [0, 1, 0], [0, 0, 1])
-        >>> robot.attach_tool(Tool(mesh, frame))
-        >>> frame_tcf = Frame((-0.309, -0.046, -0.266), (0.276, 0.926, -0.256), (0.879, -0.136, 0.456))
-        >>> robot.to_tool0_frame(frame_tcf)
-        Frame(Point(-0.363, 0.003, -0.147), Vector(0.388, -0.351, -0.852), Vector(0.276, 0.926, -0.256))
-        >>> robot.detach_tool()
-        """
-        if not self.attached_tool:
-            raise Exception("Please attach a tool first.")
-        Te = Transformation.from_frame_to_frame(self.attached_tool.frame, Frame.worldXY())
-        Tc = Transformation.from_frame(frame_tcf)
-        return Frame.from_transformation(Tc * Te)
-
-    def to_tool0_frames(self, frames_tcf):
+    def from_attached_tool_to_tool0(self, frames_tcf):
         """Converts a list of frames at the robot's tool tip (tcf frame) to frames at the robot's flange (tool0 frame) using the attached tool.
 
         Parameters
@@ -738,7 +704,7 @@ class Robot(object):
         >>> frame = Frame([0.14, 0, 0], [0, 1, 0], [0, 0, 1])
         >>> robot.attach_tool(Tool(mesh, frame))
         >>> frames_tcf = [Frame((-0.309, -0.046, -0.266), (0.276, 0.926, -0.256), (0.879, -0.136, 0.456))]
-        >>> robot.to_tool0_frames(frames_tcf)
+        >>> robot.from_attached_tool_to_tool0(frames_tcf)
         [Frame(Point(-0.363, 0.003, -0.147), Vector(0.388, -0.351, -0.852), Vector(0.276, 0.926, -0.256))]
         """
         if not self.attached_tool:
@@ -746,40 +712,7 @@ class Robot(object):
         Te = Transformation.from_frame_to_frame(self.attached_tool.frame, Frame.worldXY())
         return [Frame.from_transformation(Transformation.from_frame(f) * Te) for f in frames_tcf]
 
-    def to_tool_frame(self, frame_t0cf):
-        """Converts a frame at the robot's flange (tool0 frame) to a frame at the robot's tool tip (tcf frame) using the attached tool.
-
-        Parameters
-        ----------
-        frame_t0cf : :class:`Frame`
-            A frame (in WCF) at the robot's flange (tool0).
-
-        Returns
-        -------
-        :class:`Frame`
-            A frame (in WCF) at the robot's tool tip (tcf).
-
-        Raises
-        ------
-        Exception
-            If the end effector is not set.
-
-        Examples
-        --------
-        >>> mesh = Mesh.from_stl(compas_fab.get('planning_scene/cone.stl'))
-        >>> frame = Frame([0.14, 0, 0], [0, 1, 0], [0, 0, 1])
-        >>> robot.attach_tool(Tool(mesh, frame))
-        >>> frame_t0cf = Frame((-0.363, 0.003, -0.147), (0.388, -0.351, -0.852), (0.276, 0.926, -0.256))
-        >>> robot.to_tool_frame(frame_t0cf)
-        Frame(Point(-0.309, -0.046, -0.266), Vector(0.276, 0.926, -0.256), Vector(0.879, -0.136, 0.456))
-        """
-        if not self.attached_tool:
-            raise Exception("Please attach a tool first.")
-        Te = Transformation.from_frame_to_frame(Frame.worldXY(), self.attached_tool.frame)
-        Tc = Transformation.from_frame(frame_t0cf)
-        return Frame.from_transformation(Tc * Te)
-
-    def to_tool_frames(self, frames_t0cf):
+    def from_tool0_to_attached_tool(self, frames_t0cf):
         """Converts frames at the robot's flange (tool0 frame) to frames at the robot's tool tip (tcf frame) using the attached tool.
 
         Parameters
@@ -803,7 +736,7 @@ class Robot(object):
         >>> frame = Frame([0.14, 0, 0], [0, 1, 0], [0, 0, 1])
         >>> robot.attach_tool(Tool(mesh, frame))
         >>> frames_t0cf = [Frame((-0.363, 0.003, -0.147), (0.388, -0.351, -0.852), (0.276, 0.926, -0.256))]
-        >>> robot.to_tool_frames(frames_t0cf)
+        >>> robot.from_tool0_to_attached_tool(frames_t0cf)
         [Frame(Point(-0.309, -0.046, -0.266), Vector(0.276, 0.926, -0.256), Vector(0.879, -0.136, 0.456))]
         """
         if not self.attached_tool: