+++ CHANGELOG.rst
@@ -12,7 +12,7 @@ Unreleased
 
 **Added**
 
-* Added ``attach_tool``, ``detach_tool``, ``draw_attached_tool``, ``to_tool_frame(s)`` and ``to_tool0_frame(s)`` to ``Robot``
+* Added ``attach_tool``, ``detach_tool``, ``draw_attached_tool``, ``from_tool0_to_attached_tool`` and ``from_attached_tool_to_tool0`` to ``Robot``
 * Added ``attach_tool`` and ``detach_tool`` to ``Artist``
 * Added ``add_attached_tool`` and ``remove_attached_tool`` to ``PlanningScene``